+++ misc/update_DESCRIPTION_NEWS.R
@@ -4,12 +4,15 @@
 unlink("DESCRIPTION")
 unlink("NEWS.md")
 
+if (!requireNamespace("newsmd", quietly = TRUE)) {
+  devtools::install_github("Dschaykib/newsmd")
+}
 
 # initial files -----------------------------------------------------------
 
 # Create a new description object
 my_desc <- desc::description$new("!new")
-my_news <- newsmd()
+my_news <- newsmd::newsmd()
 
 # Set your package name
 my_desc$set("Package", "newsmd")
@@ -21,17 +24,20 @@ my_desc$del("Maintainer")
 # Set the version
 my_desc$set_version("0.0.0.9000")
 # The title of your package
-my_desc$set(Title = "creation of NEWS.md file")
+my_desc$set(Title = "Creation of NEWS.md File")
 # The description of your package
 my_desc$set(Description =
-  paste0("Adding updates (version or bulletpoints) to the NEWS.md file."))
+  paste0("Adding updates (version or bullet points) to the NEWS.md file."))
 # The urls
 my_desc$set("URL", "https://github.com/Dschaykib/newsmd")
 my_desc$set("BugReports",
             "https://github.com/Dschaykib/newsmd/issues")
 
 #Set authors
-my_desc$set("Authors@R", "person('Jakob', 'Gepp', email = 'jakob.gepp@yahoo.de', role = c('cre', 'aut'))")
+my_desc$set("Authors@R",
+            paste0("person('Jakob', 'Gepp',",
+                   "email = 'jakob.gepp@yahoo.de',",
+                   "role = c('cre', 'aut'))"))
 
 # set R version
 my_desc$set_dep("R", type = desc::dep_types[2], version = ">= 3.3.3")
@@ -53,22 +59,24 @@ my_news$add_bullet(c("adding newsmd for easier creation",
                      "adding circleci"))
 my_news$add_subtitle("Bugfix")
 my_news$add_bullet(
-  paste0("print method only shows last verion ",
+  paste0("print method only shows last version ",
          "(fix [issue #2](https://github.com/Dschaykib/newsmd/issues/2))"))
 
 
 # changing travis setup and adding lintr ----------------------------------
+
 my_desc$bump_version("minor")
 my_news$add_version(my_desc$get_version())
 my_desc$set_dep("lintr", type = desc::dep_types[3], version = "*")
 
 my_news$add_bullet(c("changing travis setup",
-                     "adding lintr"))
+                     "adding lint checks"))
 my_news$add_subtitle("Style")
-my_news$add_bullet(paste0("changing inital message"))
+my_news$add_bullet(paste0("changing initial message"))
 
 
-# changing to github action -----------------------------------------------
+# changing to GitHub action -----------------------------------------------
+
 my_desc$bump_version("minor")
 my_news$add_version(my_desc$get_version())
 # change R version dependency
@@ -76,13 +84,47 @@ my_news$add_version(my_desc$get_version())
 my_desc$set_dep("R", type = desc::dep_types[2], version = ">= 3.3")
 
 my_news$add_bullet(c("removing travis, appveyor and codecov yml",
-                     "adding github actions"))
+                     "adding GitHub actions"))
+
+
+
+# prepare for CRAN --------------------------------------------------------
+
+my_desc$bump_version("dev")
+my_news$add_version(my_desc$get_version())
+# add dependencies for vignette
+my_desc$set_dep("knitr", type = desc::dep_types[3], version = "*")
+my_desc$set_dep("rmarkdown", type = desc::dep_types[3], version = "*")
+my_desc$set_dep("desc", type = desc::dep_types[3], version = "*")
+my_desc$set(VignetteBuilder = "knitr")
+my_desc$set(Language = "en-GB")
+
+my_news$add_bullet(c("adding CRAN test and setup for release",
+                     "change test setup from ubuntu 16.04 to 18.04"))
+
+
+
+# CRAN release ------------------------------------------------------------
+
+my_desc$bump_version("patch")
+my_news$add_version(my_desc$get_version())
+# add dependencies for vignette
+my_news$add_bullet(c("first CRAN release"))
 
 # save everything ---------------------------------------------------------
-
 
 my_desc$set("Date", Sys.Date())
 my_desc$write(file = "DESCRIPTION")
 my_news$write(file = "NEWS.md")
 
+# set version number in README
+my_readme <- readLines("README.md")
+my_readme[1] <- paste0(
+  "# newsmd - ", my_desc$get_version(),
+  " <img src=\"misc/news.png\" width=170 align=\"right\" />")
+writeLines(my_readme, "README.md")
 
+# update documentation
+roxygen2::roxygenise()
+# tidy DESCRIPTON
+usethis::use_tidy_description()