+++ NEWS.md
@@ -1,9 +1,22 @@
+## version 0.3.1
+
+---
+
+- first CRAN release
+
+## version 0.3.0.9000
+
+---
+
+- adding CRAN test and setup for release
+- change test setup from ubuntu 16.04 to 18.04
+
 ## version 0.3.0
 
 ---
 
 - removing travis, appveyor and codecov yml
-- adding github actions
+- adding GitHub actions
 
 ## version 0.2.0
 
@@ -12,10 +25,10 @@
 
 ### Style
 
-- changing inital message
+- changing initial message
 
 - changing travis setup
-- adding lintr
+- adding lint checks
 
 ## version 0.1.0
 
@@ -24,7 +37,7 @@
 
 ### Bugfix
 
-- print method only shows last verion (fix [issue #2](https://github.com/Dschaykib/newsmd/issues/2))
+- print method only shows last version (fix [issue #2](https://github.com/Dschaykib/newsmd/issues/2))
 
 - adding newsmd for easier creation
 - adding circleci